Styled component refactor - Part 1
