+++ client/components/MainHeader/MainHeader.jsx
@@ -1,17 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import Stars from './Stars';
-
-const MainHeading = styled.h3`
-  font-family: ${(({ theme: { fonts } }) => `${fonts[1]}`)};
-  font-size: 26px;
-  font-weight: 300;
-  display: inline-block;
-  -webkit-font-smoothing: antialiased;
-  line-height: 42px;
-  margin: 0 6px 0 0;
-`;
+import Stars from '../Stars/Stars';
+import MainHeading from './MainHeader.styles';
 
 const MainHeader = (props) => (
   <div>
