+++ client/components/ReviewHeader/ReviewHeader.styles.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const Container = styled.div`
+  font-size: 13px;
+  color: ${(props) => props.theme.colors.middleGray};
+  margin-bottom: 6px;
+`;
+
+export default Container;
