+++ client/components/ReviewListItem/ReviewListItem.styles.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Container = styled.div`
+  padding: 6px;
+  margin: 24px 0 30px;
+`;
+
+export default Container;
