+++ client/components/ReviewPager/ReviewPager.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReviewList from './ReviewList';
-import PageList from './PageList';
+import ReviewList from '../ReviewList/ReviewList';
+import PageList from '../PageList/PageList';
 
 const ReviewPager = (props) => (
   <div>
@@ -15,12 +15,18 @@ const ReviewPager = (props) => (
 );
 
 ReviewPager.propTypes = {
-  reviewList: PropTypes.arrayOf(PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-    PropTypes.object,
-    PropTypes.array,
-  ])).isRequired,
+  reviewList: PropTypes.arrayOf(PropTypes.shape({
+    userPicture: PropTypes.string,
+    user_profile_url: PropTypes.string,
+    username: PropTypes.string,
+    date: PropTypes.string,
+    rating: PropTypes.number,
+    text: PropTypes.string,
+    reviewPicture: PropTypes.string,
+    mainImage: PropTypes.string,
+    itemName: PropTypes.string,
+    product_id: PropTypes.number,
+  })).isRequired,
   totalPages: PropTypes.number.isRequired,
   activePage: PropTypes.number.isRequired,
   handlePageClick: PropTypes.func.isRequired,
