+++ client/components/ReviewTab/ReviewTab.styles.jsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+margin-bottom: 20px;
+overflow: visible;
+width: auto;
+padding-left: 0px;
+`;
+
+export const DropdownCont = styled.div`
+justify-content: flex-end;
+display: flex;
+box-sizing: border-box;
+margin-top: 6px;
+`;
